requirements.txt
Flask==2.3.2
Flask-Login==0.6.1
mysql-connector-python==8.0.33
requests==2.28.2
python-dotenv==1.0.0
Werkzeug==2.3.5